structure/ clean code - chaged imports and names
