+++ src/middlewares/validateJWT.ts
@@ -2,30 +2,36 @@ import { Response, Request, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 import User from '../models/user';
 
-export const validateJWT = async (req: Request, res: Response, next: NextFunction) => {
-	const token = req.header('authorization');
-	if (!token) return res.status(407).send({ msg: 'You must be logged into do this.' });
+export const validateJWT = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  const token = req.header('authorization');
+  if (!token)
+    return res.status(407).send({ msg: 'You must be logged into do this.' });
 
-	try {
-		const privateKey: string | undefined = process.env.SECRETORPRIVATEKEY;
-		if (!privateKey) throw Error('Environment variables not loaded (validateJWT)');
-		// @ts-ignore
-		const { id } = jwt.verify(token, privateKey);
+  try {
+    const privateKey: string | undefined = process.env.SECRETORPRIVATEKEY;
+    if (!privateKey)
+      throw Error('Environment variables not loaded (validateJWT)');
+    // @ts-ignore
+    const { id } = jwt.verify(token, privateKey);
 
-		//Verify ID in DB
-		const user = await User.findById(id);
+    //Verify ID in DB
+    const user = await User.findById(id);
 
-		if (!user) return res.status(401).send({ msg: 'User not found' });
+    if (!user) return res.status(401).send({ msg: 'User not found' });
 
-		//Verify user state
-		if (!user.state) return res.status(401).send({ msg: 'User blocked.' });
+    //Verify user state
+    if (!user.state) return res.status(401).send({ msg: 'User blocked' });
 
-		req.user = user;
-		return next();
-	} catch (error) {
-		console.log(`Internal server error validating JWT ${error}`);
-		return res.status(500).send({ msg: 'Internal server error' });
-	}
+    req.user = user;
+    return next();
+  } catch (error) {
+    console.log(`Internal server error validating JWT ${error}`);
+    return res.status(500).send({ msg: 'Internal server error' });
+  }
 };
 
 // Vendo empanadas
